refactor: PaidSession 도메인을 Session의 서브클래스로 분리
